Moving recommend API to User Controller Changing the initial dataset
